SQLlite EF support added pluse db initialization

+++ Databases-Teamwork-Practical-Project/SantasToyFactory/SantasToyFactory.DataLayer/ToyProductionDetails.cs
@@ -4,9 +4,9 @@
 
     public class ToyProductionDetails
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
-        public int Quantity { get; set; }
+        public long Quantity { get; set; }
 
         public decimal CommissionPercent { get; set; }
     }
